LayoutTests/inspector-protocol/debugger/nested-inspectors-expected.txt

CONSOLE MESSAGE: line 1: Error: PASS: Expected exception
Checking that the inspector code can be debugged using a second inspector window.

Debugger.setPauseOnExceptions - all
PASS: Paused!
PASS: Resumed!
